搜索模块、个人信息页面手机端样式兼容
