+++ .changeset/popular-rats-vanish.md
@@ -1,5 +1,7 @@
 ---
 "@scow/lib-web": patch
+"@scow/portal-web": patch
+"@scow/mis-web": patch
 ---
 
-修复因 td 内容过长导致 table 超出页面的问题
+修复因table超出页面的问题，搜索模块、个人信息页面手机端样式兼容
